+++ database/migrations/0003_01_04_000001_create_transactions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
             $table->string('reference')->unique()->index();
             $table->string('payment_type')->nullable()->index();
             $table->string('currency')->nullable()->index();
-            $table->unsignedBigInteger('amount')->nullable()->index();
+            $table->decimal('amount', 26, 8)->nullable()->index();
             $table->string('status')->nullable()->index();
             $table->string('external_id')->nullable()->index();
             $table->mediumText('data')->nullable();
